+++ R/ReadCodeChunks.R
@@ -25,14 +25,14 @@
 #' @export
 #'
 #' @examples
-#' path <- system.file("doc", "knitr-intro.Rmd", package = "knitr")
-#' chunks <- ReadCodeChunks(path)
+#' file <- system.file("misc", "knitr-markdown.Rmd", package = "inlmisc")
+#' chunks <- ReadCodeChunks(file)
 #'
 #' attr(chunks, "path")
 #' names(chunks)
-#' chunks[["show-off"]]
+#' chunks[["named-chunk-2"]]
 #'
-#' eval(parse(text = unlist(chunks[c("show-off", "graphics")])))
+#' eval(parse(text = unlist(chunks[c("unnamed-chunk-3", "named-chunk-4")])))
 #'
 
 ReadCodeChunks <- function(path) {